Update hacs.json
bump version to v4.0.0..

a1.0.0 was basic game play and card layout.
a1.0.1 bug hunting and squashing
b1.0.0 change of plan on how to load and run the game.
b1.0.5 find out the js would not load correctly on test server.
b2.0.0 new codebase.
b3.0.0 what was working on test server would not work in real world server
v1.0.0 major backend code re write and move from bad idea to a good idea on how to run the game.
v2.0.0 another backend change and move from a dashboard button to launch the game to a side menu option.
v3.0.0 lots of code bugs squashed. the biggest was failed auth of the long-lived token but it was my real world ha that was not configured correctly and took 2 days to nail it down.
v4.0.0 the initial release code that works.. backend code and auth is 100% working. frontend code is 100% working. now its time to tweak and polish with little version bumps when needed.