Update wordplay_game.html
include the game audio in the cleanest way possible and it keeps the audio away from the game card code completely. so changing the audio is easier and does not require a complete wordplay_game.html re write. the cleanest way i could think to do it but is an untested idea that might backfire on me.