Add files via upload
Split the main wordplay_game.html in to small parts so code is more user friendly and a hell of a lot easier to maintain.  1000+ lines of javascript has been split 6 files. the main html has gone from 1866 lines of code down to 89 lines of code.. Life should be easier now.